Structure for multiple backends.
