+++ acme_diags/plotset5/set5_diags_GPCP.json
@@ -10,6 +10,7 @@
       "region": ["global"],
       "test_units": "mm/day",
       "diff_units": "mm/day",
+      "backend":"vcs",
       "test_levels":[0, 0.2, 0.5, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 12, 14, 17],
       "reference_levels":[0, 0.2, 0.5, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 12, 14, 17],
       "diff_levels":[-6, -5, -4, -3, -2, -1, -0.5, 0, 0.5, 1, 2, 3, 4, 5, 6],
